% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vertices.R
\name{check_vertices}
\alias{check_vertices}
\title{Double-checking the amount of vertices}
\usage{
check_vertices(g, no_v)
}
\arguments{
\item{g}{Network as igraph object}

\item{no_v}{number of vertices}
}
\description{
Double-checks the amount of vertices in the network if they are equal to the amount in the interaction matrix. Otherwise the missing vertices will be added to the graph.
}
