% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_tm.R
\name{deg_tm}
\alias{deg_tm}
\title{Degree centrality in a two-mode network}
\usage{
deg_tm(net, measure = c("degree", "output"), nn)
}
\arguments{
\item{net}{see \code{\link{degree_tm}}}

\item{measure}{see \code{\link{degree_tm}}}

\item{nn}{Number of actions or environmental components, respectively.}
}
\description{
Modified function of the package tnet. Every node is now considered in the returned data frame.
}
\seealso{
\code{\link{degree_tm}}
}
