% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_w.R
\name{deg_w}
\alias{deg_w}
\title{Degree centrality in a weighted network}
\usage{
deg_w(net, measure = c("degree", "output"), type = "out", alpha = 1, nn)
}
\arguments{
\item{net}{see \code{\link{degree_w}}}

\item{measure}{see \code{\link{degree_w}}}

\item{type}{see \code{\link{degree_w}}}

\item{alpha}{see \code{\link{degree_w}}}

\item{nn}{Number of actions or environmental components, respectively.}
}
\description{
Modified function of package tnet. Like for the two-mode, every node is now considered in the returned data frame.
}
\seealso{
\code{\link{degree_w}}
}
