% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eastkarun.R
\docType{data}
\name{eastkarun}
\alias{eastkarun}
\title{Interaction matrix for the sugarcane planation at the East Karun river, Iran}
\format{Matrix}
\usage{
data(eastkarun)
}
\description{
Matrix which contains the interactions derived from the technical report of the Environmental Impact Assessment (EIA)
}
\references{
Nahvi, A., Daghighi, A., Nazif, S., 2017. The environmental impact assessment of drainage systems: a case study of the Karun river sugarcane development project. Archives of Agronomy and Soil Science: 1-11. DOI:10.1080/03650340.2017.1340641
}
\keyword{datasets}
