% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kladovo.R
\docType{data}
\name{kladovo}
\alias{kladovo}
\title{Interaction matrix for the wind farms at Kladovo, Serbia}
\format{Matrix which contains the interactions derived from the technical report of the Environmental Impact Assessment (EIA)}
\usage{
data(kladovo)
}
\description{
Matrix which contains the interactions derived from the technical report of the Environmental Impact Assessment (EIA)
}
\references{
Josimovic, B., Petric, J., Milijic, S., 2014. The Use of the leopold matrix in carrying out the eia for wind farms in Serbia. Energy and Environment Research, 4(1): 43. DOI:10.5539/eer.v4n1p43
}
\keyword{datasets}
