% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallahala.R
\docType{data}
\name{tallahala}
\alias{tallahala}
\title{Interaction matrix for the Tallahala dam at the Mississipi, USA}
\format{Matrix}
\usage{
data(tallahala)
}
\description{
Matrix which contains the interactions derived from the technical report of the Environmental Impact Assessment (EIA)
}
\keyword{datasets}
