% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visstats.R
\name{visstats}
\alias{visstats}
\title{Visualization of the network and the basic network statistic}
\usage{
visstats(mat.int, negative.values = TRUE, vcat, vcat.env.comps, vcat.actions)
}
\arguments{
\item{mat.int}{Interaction matrix as matrix}

\item{negative.values}{Defines if interaction matrix contains negative interactions. Default is set to TRUE.}

\item{vcat}{A vector which contains the labels of the vertice categories.}

\item{vcat.env.comps}{A vector which contains the vertice categories of the environmental components.}

\item{vcat.actions}{A vector which contains the vertice categories of the actions.}
}
\value{
Two data frames. \code{bas.stats} contains the results of the basic network statistic and \code{deg} includes the weighted vertice degrees and mean degrees.
}
\description{
The bipartite network from the interaction matrix is build, visualized and the basic network statistics are computed
}
\details{
...
}
\examples{
data('tallahala')
vcat <- c("Water", "Land", "Biology", "Socioeconomy", "Infrastructure")
vcat.env.comps <- c(1,1,1,1,2,2,1,1,2,2,3,3,3,3,3,3,3,4,4,4,4)
vcat.actions <- c(1,1,1,5,5,5,1,3,3,5)
visstats(tallahala, negative.values = TRUE, vcat = vcat, vcat.env.comps = vcat.env.comps,
vcat.actions = vcat.actions)

data('eastkarun')
vcat <- c("Water", "Land", "Biology", "Socioeconomy")
vcat.env.comps <- c(1,2,2,2,2,1,1,3,3,3,3,3,3,4,4,4,4,4,4,4)
vcat.actions <- c(1,1,1,3,1)
visstats(eastkarun, negative.values = TRUE, vcat = vcat, vcat.env.comps = vcat.env.comps,
vcat.actions = vcat.actions)

data('kladovo')
vcat <- c("Water", "Land", "Biology", "Socioeconomy", "Infrastructure")
vcat.env.comps <- c(1,2,2,2,2,2,3,3,3,3,3,2,2,4,4,4)
vcat.actions <- c(5,5,5,5,5,5,5,5,5)
visstats(kladovo, negative.values = FALSE, vcat = vcat, vcat.env.comps = vcat.env.comps,
vcat.actions = vcat.actions)

}
\seealso{
\code{\link{assortativity_nominal}}, \code{\link{degree_tm}}, \code{\link{degree_w}}, \code{\link{distance_tm}}, \code{\link{distance_w}}, \code{\link{clustering_tm}}
}
\author{
Robin Schwemmle \email{robin.schwemmle@venus.uni-freiburg.de}
}
